backend/requirements.txt
alembic==1.14.0
blinker==1.9.0
boto3==1.35.60
botocore==1.35.60
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-translate==3.19.0
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==2.0.36
starkbank-ecdsa==2.2.0
stripe==11.3.0
timeout-decorator==0.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
